Converted Factorials.computeRecursive() to BigInteger return.

+++ src/test/java/edu/cnm/deepdive/FactorialsTest.java
@@ -3,6 +3,7 @@ package edu.cnm.deepdive;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.math.BigInteger;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
@@ -12,7 +13,7 @@ class FactorialsTest {
   @DisplayName("computeRecursive(non-exceptions)")
   @ParameterizedTest(name = "[{index}] Asserting computeRecursive({0}) == {1}.")
   @CsvFileSource(resources = "factorials-test-non-exception.csv", numLinesToSkip = 1)
-  void computeRecursive_nonException(int input, long expected) {
+  void computeRecursive_nonException(int input, BigInteger expected) {
     assertEquals(expected, Factorials.computeRecursive(input));
   }
